1、实现群内插件开关功能 2、实现群内用户禁用功能

+++ model/database/admin.go
@@ -1,8 +1,8 @@
 package database
 
 type Admin struct {
-	Id int `xorm:"id autoincr"`
-	QQ int `xorm:"qq"`
+	Id int   `xorm:"id autoincr"`
+	QQ int64 `xorm:"qq"`
 }
 
 func (Admin) TableName() string {
